requirements.txt
autopep8==2.0.1
charset-normalizer==2.1.1
idna==3.4
kafka-python==2.0.2
oauthlib==3.2.2
pycodestyle==2.10.0
python-dotenv==0.21.0
requests==2.28.1
requests-oauthlib==1.3.1
tomli==2.0.1
tweepy==4.12.1
urllib3==1.26.13
wincertstore==0.2
